requirements.txt
django==5.2.7
djangorestframework==3.16.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
